feat(#90) 매너 농도 계산&거래후기 중복 작성 방지

+++ src/main/java/com/ssd/sthub/dto/review/ReviewRequestDTO.java
@@ -14,10 +14,12 @@ public class ReviewRequestDTO {
     @Getter
     public static class Request {
         private Long purchaseId;
+
         @NotNull(message = "거래 후기 점수를 매겨 주세요.")
         @Min(value = 1, message = "점수는 1점 이상이어야 합니다.")
         @Max(value = 5, message = "점수는 5점 이하이어야 합니다.")
-        private Integer rating;
+        private Double rating;
+
         private int tag1;
         private int tag2;
         private int tag3;
@@ -25,5 +27,4 @@ public class ReviewRequestDTO {
         private int tag5;
         private int tag6;
     }
-
 }
